\name{SD-package}
\alias{SD-package}
\alias{SD}
\docType{package}
\title{
\packageTitle{SD}
}
\description{
\packageDescription{SD}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SD}

Maintainer: \packageMaintainer{SD}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SD)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SD}}
%\section{Documentation Index}{\packageIndices{SD}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
