% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Standard Deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
the standard deviation
}
\description{
Standard Deviation
}
\examples{
standard_deviation(c(1,4,7,2,8,4,9,1))
}

