% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{standard_error}
\alias{standard_error}
\title{Standard Error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
the standard error
}
\description{
Standard Error
}
\examples{
standard_error(c(1,4,7,2,8,4,9,1))
}

